chore(apps/api): refactor GetFuseLeveragedTokenSwapsBySymbol
