+++ apps/api/src/controllers/flt/index.ts
@@ -1,9 +1,11 @@
 import GetFuseLeveragedTokenBySymbol from "./GetFuseLeveragedTokenBySymbol";
 import GetFuseLeveragedTokenUserPositionById from "./GetFuseLeveragedTokenUserPositionById";
+import GetFuseLeveragedTokenSwapsBySymbol from "./GetFuseLeveragedTokenSwapsBySymbol";
 
 const controllers = {
     GetFuseLeveragedTokenBySymbol,
     GetFuseLeveragedTokenUserPositionById,
+    GetFuseLeveragedTokenSwapsBySymbol,
 };
 
 export default controllers;
