+++ apps/api/src/controllers/flts.ts
@@ -58,41 +58,6 @@ async function GetFuseLeveragedTokenChartsBySymbol(
 }
 
 /**
- * GetFuseLeveragedTokenSwapsBySymbol return swap activities
- */
-async function GetFuseLeveragedTokenSwapsBySymbol(
-    req: Request,
-    res: Response
-) {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(404).json({ errors: errors.array() });
-    }
-    try {
-        const swaps = await fltsService.getFuseLeveragedTokenSwapsBySymbol(
-            req.params.chainId as unknown as ChainId,
-            req.params.symbol,
-            req.query.userAddress as unknown as string
-        );
-        if (swaps == undefined) {
-            return res.status(404).json({
-                errors: [
-                    {
-                        location: "params",
-                        msg: "symbol not found",
-                        param: "symbol",
-                        value: req.params.symbol,
-                    },
-                ],
-            });
-        }
-        return res.status(200).json(swaps);
-    } catch (e) {
-        return res.status(500).json({ error: e });
-    }
-}
-
-/**
  * GetFuseLeveragedTokenBackingsBySymbol return backings history
  */
 async function GetFuseLeveragedTokenBackingsBySymbol(
@@ -132,7 +97,6 @@ const flts = {
     GetFuseLeveragedTokensByChainIdValidation,
     GetFuseLeveragedTokensByChainId,
     GetFuseLeveragedTokenChartsBySymbol,
-    GetFuseLeveragedTokenSwapsBySymbol,
     GetFuseLeveragedTokenBackingsBySymbol,
 };
