+++ apps/api/src/services/flts.ts
@@ -72,48 +72,6 @@ const queryFuseLeveragedTokenChartsBySymbol = gql`
     }
 `;
 
-const queryFuseLeveragedTokenSwapsBySymbol = gql`
-    query getFuseLeveragedTokenSwaps($symbol: String, $user: String) {
-        flts: flts(where: { symbol: $symbol }) {
-            swaps(first: 100, orderBy: timestamp, orderDirection: desc) {
-                ...swapInfo
-            }
-        }
-        user: flts(where: { symbol: $symbol }) {
-            swaps(
-                first: 100
-                orderBy: timestamp
-                orderDirection: desc
-                where: { user: $user }
-            ) {
-                ...swapInfo
-            }
-        }
-    }
-
-    fragment swapInfo on Swap {
-        timestamp
-        transaction {
-            id
-        }
-        amountIn
-        amountInUSD
-        amountOut
-        amountOutUSD
-        user {
-            id
-        }
-        tokenIn {
-            name
-            symbol
-        }
-        tokenOut {
-            name
-            symbol
-        }
-    }
-`;
-
 const queryFuseLeveragedTokenBackingsBySymbol = gql`
     query getFuseLeveagedTokenBackings($symbol: String) {
         flts(where: { symbol: $symbol }) {
@@ -323,55 +281,6 @@ export async function getFuseLeveragedTokenChartsBySymbol(
 }
 
 /**
- * Get Fuse Leveraged Token swaps activity
- */
-export async function getFuseLeveragedTokenSwapsBySymbol(
-    chainId: ChainId,
-    fltSymbol: string,
-    userAddress: string | undefined
-): Promise<FuseLeveragedTokenSwaps | undefined> {
-    // Get data from the graph
-    const endpoint = getGraphEndpointByChainId(chainId);
-    const filter = fltSymbol.toUpperCase();
-    const data = await grequest(
-        endpoint,
-        queryFuseLeveragedTokenSwapsBySymbol,
-        {
-            symbol: filter,
-            user: userAddress,
-        }
-    );
-    if (data.flts.length == 0) return undefined;
-    const flt = data.flts[0].swaps.map((swap: any) => {
-        return {
-            timestamp: parseInt(swap.timestamp),
-            hash: swap.transaction.id,
-            user: swap.user.id,
-            tokenIn: swap.tokenIn,
-            amountIn: parseFloat(swap.amountIn),
-            amountInUSD: parseFloat(swap.amountInUSD),
-            tokenOut: swap.tokenOut,
-            amountOut: parseFloat(swap.amountOut),
-            amountOutUSD: parseFloat(swap.amountOutUSD),
-        };
-    });
-    const user = data.user[0].swaps.map((swap: any) => {
-        return {
-            timestamp: parseInt(swap.timestamp),
-            hash: swap.transaction.id,
-            user: swap.user.id,
-            tokenIn: swap.tokenIn,
-            amountIn: parseFloat(swap.amountIn),
-            amountInUSD: parseFloat(swap.amountInUSD),
-            tokenOut: swap.tokenOut,
-            amountOut: parseFloat(swap.amountOut),
-            amountOutUSD: parseFloat(swap.amountOutUSD),
-        };
-    });
-    return { flt, user };
-}
-
-/**
  * Get Fuse Leveraged Token historical backings
  */
 export async function getFuseLeveragedTokenBackingsBySymbol(
@@ -403,7 +312,6 @@ export async function getFuseLeveragedTokenBackingsBySymbol(
 const flts = {
     getFuseLeveragedTokensByChainId,
     getFuseLeveragedTokenChartsBySymbol,
-    getFuseLeveragedTokenSwapsBySymbol,
     getFuseLeveragedTokenBackingsBySymbol,
 };
