+++ apps/api/tests/services/GetFuseLeveragedTokenSwapsBySymbol.spec.ts
@@ -1,12 +1,13 @@
-import { getFuseLeveragedTokenSwapsBySymbol } from "../flts";
 import { ChainId } from "@risedle/types";
 
-describe("getFuseLeveragedTokenSwapsBySymbol", () => {
+import flt from "../../src/services/flt";
+
+describe("GetFuseLeveragedTokenSwapsBySymbol", () => {
     describe("given random chainId", () => {
         it("should throw an error", async () => {
             expect.assertions(1);
             try {
-                await getFuseLeveragedTokenSwapsBySymbol(
+                await flt.GetFuseLeveragedTokenSwapsBySymbol(
                     1234,
                     "AAA",
                     undefined
@@ -20,7 +21,7 @@ describe("getFuseLeveragedTokenSwapsBySymbol", () => {
     describe("given BSC as chainId", () => {
         describe("given random symbol", () => {
             it("should return undefined", async () => {
-                const swaps = await getFuseLeveragedTokenSwapsBySymbol(
+                const swaps = await flt.GetFuseLeveragedTokenSwapsBySymbol(
                     56,
                     "AAA",
                     undefined
@@ -30,8 +31,8 @@ describe("getFuseLeveragedTokenSwapsBySymbol", () => {
         });
 
         describe("given BNBRISE without userAddress", () => {
-            it("should return array of FuseLeveragedTokenPrice", async () => {
-                let swaps = await getFuseLeveragedTokenSwapsBySymbol(
+            it("should return array of FuseLeveragedTokenSwap", async () => {
+                let swaps = await flt.GetFuseLeveragedTokenSwapsBySymbol(
                     56,
                     "BNBRISE",
                     undefined
@@ -61,10 +62,56 @@ describe("getFuseLeveragedTokenSwapsBySymbol", () => {
             });
         });
 
-        describe("given BNBRISE with userAddress", () => {
-            it("should return array of FuseLeveragedTokenPrice", async () => {
+        describe("given BNBRISE with lowercased userAddress", () => {
+            it("should return array of FuseLeveragedTokenSwap", async () => {
                 const addy = "0x1418be4753a22b69b613fa8b8144d856c023d46b";
-                let swaps = await getFuseLeveragedTokenSwapsBySymbol(
+                let swaps = await flt.GetFuseLeveragedTokenSwapsBySymbol(
+                    56,
+                    "BNBRISE",
+                    addy
+                );
+                expect(swaps?.flt.length).toBeGreaterThanOrEqual(1);
+                expect(swaps?.user.length).toBeGreaterThanOrEqual(1);
+
+                // Make sure it returns array of FuseLeveragedTokenSwap
+                const fltSwaps = swaps?.flt!;
+                expect(fltSwaps[0].timestamp).toBeGreaterThan(10000);
+                expect(typeof fltSwaps[0].hash).toBe("string");
+                expect(fltSwaps[0].hash.length).toBe(66);
+                expect(typeof fltSwaps[0].user).toBe("string");
+                expect(fltSwaps[0].user.length).toBe(42);
+                expect(typeof fltSwaps[0].tokenIn.name).toBe("string");
+                expect(typeof fltSwaps[0].tokenIn.symbol).toBe("string");
+                expect(typeof fltSwaps[0].amountIn).toBe("number");
+                expect(typeof fltSwaps[0].amountInUSD).toBe("number");
+                expect(typeof fltSwaps[0].tokenOut.name).toBe("string");
+                expect(typeof fltSwaps[0].tokenOut.symbol).toBe("string");
+                expect(typeof fltSwaps[0].amountOut).toBe("number");
+                expect(typeof fltSwaps[0].amountOutUSD).toBe("number");
+
+                // Check user swaps; Make sure it returns Swap data
+                const userSwaps = swaps?.user!;
+                expect(userSwaps[0].timestamp).toBeGreaterThan(10000);
+                expect(typeof userSwaps[0].hash).toBe("string");
+                expect(userSwaps[0].hash.length).toBe(66);
+                expect(typeof userSwaps[0].user).toBe("string");
+                expect(userSwaps[0].user.length).toBe(42);
+                expect(userSwaps[0].user).toBe(addy);
+                expect(typeof userSwaps[0].tokenIn.name).toBe("string");
+                expect(typeof userSwaps[0].tokenIn.symbol).toBe("string");
+                expect(typeof userSwaps[0].amountIn).toBe("number");
+                expect(typeof userSwaps[0].amountInUSD).toBe("number");
+                expect(typeof userSwaps[0].tokenOut.name).toBe("string");
+                expect(typeof userSwaps[0].tokenOut.symbol).toBe("string");
+                expect(typeof userSwaps[0].amountOut).toBe("number");
+                expect(typeof userSwaps[0].amountOutUSD).toBe("number");
+            });
+        });
+
+        describe("given BNBRISE with checksummed userAddress", () => {
+            it("should return array of FuseLeveragedTokenSwap", async () => {
+                const addy = "0x1418bE4753a22b69b613fA8B8144D856C023D46B";
+                let swaps = await flt.GetFuseLeveragedTokenSwapsBySymbol(
                     56,
                     "BNBRISE",
                     addy
